Improved auth
