+++ .idea/misc.xml
@@ -3,6 +3,22 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/activity_main.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/activity_start.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/content_main.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_add_member.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_dashboard.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_login.xml" value="0.22" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_members.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_register.xml" value="0.4317708333333333" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_room_view.xml" value="0.4317708333333333" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/fragment_settings.xml" value="0.4317708333333333" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/include_menu_buttons.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/include_navbar_buttons.xml" value="0.24635416666666668" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/member_grid_item.xml" value="0.4317708333333333" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/sample_item_member.xml" value="0.41770833333333335" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/settings_grid_item.xml" value="0.4317708333333333" />
+        <entry key="..\:/TBZ/MyHome/MyHome-Mobile/app/src/main/res/layout/top_toolbar.xml" value="0.4317708333333333" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.3390625" />
         <entry key="app/src/main/res/drawable/background_tint_light.xml" value="0.3390625" />
         <entry key="app/src/main/res/drawable/background_tint_night.xml" value="0.3390625" />
@@ -24,11 +40,10 @@
         <entry key="app/src/main/res/layout/fragment_settings.xml" value="0.18385416666666668" />
         <entry key="app/src/main/res/layout/include_menu_buttons.xml" value="0.18385416666666668" />
         <entry key="app/src/main/res/layout/navigation_buttons.xml" value="0.3385416666666667" />
-        <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.3546875" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="Embedded JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
