+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".MainActivity">
+        tools:context=".Application">
 
     <include layout="@layout/content_main" />
